refactor(ingest/graph): Factor out filter logic
